create UE model
- refactor the cell handling
