+++ src/importer/md2model/cellHandlers/utils.js
@@ -9,30 +9,8 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-export function getBlockName(node) {
-  return node.properties.className[0];
-}
-
-export function toCamelCase(str) {
-  return str
-    .toLowerCase()
-    .replace(/[^a-zA-Z0-9]+(.)/g, (match, chr) => chr.toUpperCase());
-}
-
-export function getModelDefinition(blockName) {
-  return {
-    title: blockName,
-    id: blockName,
-    plugins: {
-      xwalk: {
-        page: {
-          resourceType: 'core/franklin/components/block/v1/block',
-          template: {
-            name: blockName,
-            model: blockName,
-          },
-        },
-      },
-    },
-  };
+export function firstChildIsText(node) {
+  return node.children.length > 0
+    && node.children[0].type === 'text'
+    && node.children[0].value.trim().length > 0;
 }
